lambda_functions/master/requirements.txt
pymysql
python-dotenv
boto3
botocore
